requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.0
numpy>=1.24.0
matplotlib>=3.8.0
openai>=1.6.0
python-dotenv>=1.0.0
tomli>=2.0.0
websockets>=12.0
